Introduce prepare_stream and stream using cursors
Adds prepare_stream and stream that stream results using new callbacks
handle_open/4, handle_fetch/4, handle_close/4.

The prepare phase of prepare_stream follows the same rules as
prepare_execute: close query on describe or encode error. Other errors
do not cause a close.

For the cursor `handle_close/4` is not called if `handle_fetch/4`
returns a :done tuple, it is assumed `handle_fetch/4` did any required
cleanup. Otherwise cursor is closed.

The cursor is a constant throughout the stream because it may not be
possible to recover a change in cursor if an error occurs. As a query
should not be executed in `handle_open/4`, but started in
`handle_fetch/4, cursors may need to be tracked in the callback module.