Strictly revoke ownership of allowed processes on proxy checkin
If a process is allowed on a connection/proxy that is checked in the
process is marked revoked and requires explicit checkout/allow to
regain an allowance on a process.

Note that shared mode is as an explicit allow, and so delays the revoking
until the shared mode ends. If shared mode is enter multiple times in a
row the revoke is delayed each time. Also caller: pid will continue to work.

In practice this means the only new time a new failure occurs is when a
process is allowed on a connection/proxy that gets checked in, it can not
gain a new connection in automatic mode without explicit checkout/allow.
In the same situation in manual mode the error warns about revoking but
an explicit checkout/allow was required.