Add built in connection pool (#108)
First step towards a faster/ singe pool written specifically for DBConnection. This pool uses ETS tables to store connection state and moves ownership of the connection around to avoid tracking state internally. The pool is faster because interaction between the clients and the pool requires only a single message roundtrip to checkout, and a single cast to return. Thus beating or matching the efficiency of our existing pools in each case.