Make sure idle_interval considers the current time
Imagine we have two connections, they were checked in at
timestamps 500 and 600 (ms).

At moment 0, we emit the first interval check. We will check
at interval 1000 if there is any timestamp older than 0. When
interval 1000 arrives, we see that 500 is more than 0, so we
don't ping anything and we schedule another check. Prior to
this patch, the new interval wouldd check if there is a connection
older than 500 - the last queued in time - which would cause it
only to ping the connection at timestamp 500, while both should
have been dropped.

With this patch, the following happens. At moment 0, we emit the
first interval check. We will check at interval 1000 if there is
any timestamp older than 0. When interval 1000 arrives, we see
that 500 is more than 0, so we don't ping anything and we schedule
another check to see if there is a connection older than 1000, which
is the current time. This will cause both connections 500 and 600
to be pinged.

The new behaviour can be achieved by simply removing 2 lines of
code. The remaining changes were a clean-up to make `start_idle`
effectively run on rolling windows.

While I able to verify the fix in tests, I could not effectively
write a test for it. Prior to this patch, I could see that 99% of
the time we would never ping two connections at the same time.
With the fix, we ping all connections 99% of the time. Unfortunately,
the 1% means the test could have false negatives.

Closes #216